Initialize repo.
`weather_cli` is a simple toy project extending upon a tutorial program
indended to show off working with `click`. This program provides a
simple CLI for pulling current weather conditions for a specified
location. This initial revision represents the tutorial program with
some minor changes to output formatting.

This code was put together based on the following guides:

  * https://dbader.org/blog/python-commandline-tools-with-click
  * https://dbader.org/blog/mastering-click-advanced-python-command-line-apps