Add `setuptools` support
Finally figured out how to use `setuptools` for ease of installation and
use. This allows the user to install and set up this program with a
single command. Also removes the need for the user to manage the virtual
environment, or determine dependencies. Should allow for a more simple
deploy & use step.