Add outut type option
The output type option allows the user to choose the amount of weather
data to be displayed. Currently the default, and only implemented more
is of the full output. Not technically the full data, but full in the
sense that it is the same data returned in the initial revision.

This flag will likely be modified into a series of boolean flags which
toggle specific data to be displayed or not. Boolean flags will allow
the user full control over the output to display, useful when
interfacing this program with others.

This change also begins to make our weather api functions slightly more
generic in that they return whole sets of data instead of particular
pieces based on the command run. This should allow us more flexibility
when extending the functionality.