Pretty printing flag
The pretty printing flag, when enabled, will add some extra 'flavor
text' or formatting to the output. This is useful for running in a
terminal, where you may want to see the weather displayed nicely.

By default, this flag is disabled (maybe switch that?) and the output
will only include the values specified in the arguments. This is useful
when chaining this program's output into another, or saving it into a
log file.

While adding this flag, I noticed an output formatting issue. Depending
on the options selected, there may or may not be leading or trailing
commas where they should not be. Should be a relatively simple fix. Just
move some things around, maybe make a few checks in other places. Simply
wanted to make a note of this here so it is not forgotten.