Add flag for wind speed/direction
When activated the `--wind/-w` flag will display the current wind speed
and direction in the output. Using the same basic idiom explained to me
for the temperature units made this a fairly simple addition.

Coming soon to a commit near you: display the wind direction as a
cardinal direction instead of specific degrees. A few `if/else`
statements should do the trick. But there may be a better way.