Add `degrees_to_cardinal` helper function
This function will accept a number of degrees, and return a string
representing the cardinal direction. Simple enough, though I did search
around for a more efficient solution than rows of `if/else` statements.

The original function was take from a GitHub Gist, and has also been
updated and forked into a new Gist as the original was not updated to
include the changes in some comments.

Original Gist:
  - https://gist.github.com/RobertSudwarts/acf8df23a16afdb5837f
My Fork:
  - https://gist.github.com/ncdulo/7861c85518c3f1091ea36d4eb62b8593