Update install instructions
Continuing my eternal confusing with the `ls` command, I realized the original command was not correct. To properly create the symlink, the full path must be used. Update the `ln` command to include `pwd` as a means of grabbing the proper, full path to the executable.

Alternatively, the full path can just be directly specified and the `pwd` bit omitted from that step.