Remove manual API key install instruction
Further testing shows that manually installing the API key by running an `echo API_KEY...` command does not work as intended. Seems to be related to the line endings. Having not totally thought this one through, I am omitting the manual install instruction in favor of using the built-in `config` command. The `config` command works as intended and should be the standard method used to install the API key.