Basic exception handling for API request
Bring in some basic exception handling for potential exceptions raised
by `requests` module. This includes HTTP response errors, connection
errors, and connection timeouts. This is by no means an exhaustive list
of all possible errors to handle, but it should handle the most common
ones. At least regarding the `requests` module.

Even in the case of an exception being raised, there is still error
information within the `Response` object returned by `current_weather`.
We may use that text in favor of the text returned by the `requests`
exception as it may be more informative and specific to our application.