Create CODE_OF_CONDUCT.md
I mostly just wanted this to fill in the check marks on the `Insights->Community` page for this repo. Play nice.

Yes, I realize this is a slightly out of date version of the Contributor Covenant. I read into version 2.0 but I did not like the section with the 'enforcement actions'. Sticking to the tried & true.