Update `--pretty` flag to function as a choice.
Using templates to pretty-print our output means we can have multiple to
choose from. Therefore it makes sense we should update the `--pretty` flag
to function as a choice for output format. In this state, the options do
nothing at all. Check ahead a few commits for the actual implementation.

I am undecided on the names for each pretty mode. Partly because I am
uncertain about which pretty modes I will add. For now, this seems like
a sensible start. The default `short` choice to prefer a sensible amount
of output. I'm certain these will be expanded upon in the future.