Implement proper output templates via `--pretty`.
This commit brings in the functionality for our new `--pretty` output
formats. We start off by defining a series of output templates, and pick
one based on the user specied arguments.

Then, as a slight modification to our old string concatenation Hell, we
build up a dictionary of the current conditions. That dictionary gets
passed into `Template.safe_substitute()` for string replacement before
finally displaying the data to the user.

We use `safe_substitute()` to allow missing template keys to be skipped
over without an error. Goes in hand with our key checks in the code
assembling the dictionary.