Add demo output
I had posted this demo output onto Discord to show off some of the ways arguments can be passed in. As well as to show off the new output formats feature. Figured I would add it in to the README for a simple demo of the program output in an easy to access location.