Handle disabled flag cases, partially.
When a certain flag for output is disabled, we want to skip over it
silently. The previous implementation would leave template placeholders
in the output when a flag was disabled. The code has been modified so
that disabled flags have their entry in the `current_conditions`
dictionary set to an empty string.

This works in that we silently skip over disabled flags now. However, we
still have strange gaps in the output format where the parameter for
that flag would be in the output. This makes sense, our templates were
written with all flags enabled in mind.

Think through the output formats and modify to suit, or update the logic
used to assemble them.